requirements.txt
streamlit>=1.30.0
streamlit-sortables>=0.3.0
python-pptx>=0.6.21
openpyxl>=3.1.0
pandas>=2.0.0
pyyaml>=6.0
xlwings>=0.30.0
lxml>=4.9.0
Pillow>=10.0.0
